archive styling and show select
