+++ frontend/src/metabase/components/ArchivedItem.jsx
@@ -3,11 +3,12 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { t } from "c-3po";
-import Icon from "metabase/components/Icon";
-import Tooltip from "metabase/components/Tooltip";
+
 import CheckBox from "metabase/components/CheckBox.jsx";
-import { Box } from "rebass";
-import cx from "classnames";
+import Icon from "metabase/components/Icon";
+import IconWrapper from "metabase/components/IconWrapper"
+import Swapper from "metabase/components/Swapper"
+import Tooltip from "metabase/components/Tooltip";
 
 const ArchivedItem = ({
   name,
@@ -19,18 +20,20 @@ const ArchivedItem = ({
 
   selected,
   onToggleSelected,
+  showSelect,
 }) => (
   <div className="flex align-center p2 hover-parent hover--visibility border-bottom bg-grey-0-hover">
-    <Box className="hover-parent hover--visibility">
-      <Box className="hover-child hover-child--hiden">
-        <Icon name={icon} className="mr2" style={{ color: color }} size={20} />
-      </Box>
-      {onToggleSelected && (
-        <Box className={cx({ "hover-child": !selected })}>
+    <IconWrapper p={1} mr={1} align='center' justify='center'>
+      <Swapper
+        startSwapped={showSelect}
+        defaultElement={
+          <Icon name={icon} color={color} />
+        }
+        swappedElement={
           <CheckBox checked={selected} onChange={onToggleSelected} />
-        </Box>
-      )}
-    </Box>
+        }
+      />
+    </IconWrapper>
     {name}
     {isAdmin && (
       <Tooltip tooltip={t`Unarchive this ${type}`}>
