+++ frontend/src/metabase/components/IconWrapper.jsx
@@ -1,0 +1,8 @@
+import { Flex } from "rebass"
+
+const IconWrapper = Flex.extend`
+  background: #f4f5f6;
+  border-radius: 6px;
+`;
+
+export default IconWrapper
